feat: create a shared Counter component; chore
