+++ packages/shared/components/Counter/Counter.tsx
@@ -1,0 +1,24 @@
+import { component$ } from '@builder.io/qwik';
+import type { CounterProps } from './Counter.type';
+
+export default component$(({ label, onClick$, state }: CounterProps) => {
+	return (
+		<button
+			style={{
+				border: '0 solid #e2e8f0',
+				marginTop: '10px',
+				backgroundColor: 'rgb(246, 179, 82)',
+				borderRadius: '.25rem',
+				fontWeight: '700',
+				padding: '.5rem 1rem .5rem 1rem',
+				color: 'rgb(24, 24, 24)',
+			}}
+			onClick$={() => {
+				state.increment();
+				onClick$ && onClick$();
+			}}
+		>
+			{label} {state.counter}
+		</button>
+	);
+});
