+++ packages/shared/components/Counter/Counter.type.ts
@@ -1,0 +1,9 @@
+import type { QRL } from "@builder.io/qwik";
+
+export interface CounterProps {
+  label: string;
+  onClick$?: QRL<() => void>;
+  state: CounterStore;
+}
+
+export type CounterStore = { counter: number; increment: QRL<(this: CounterStore) => void> };
