+++ packages/shared/components/RemoreMfe/RemoteMfe.types.ts
@@ -1,0 +1,10 @@
+export interface RemoteData {
+	name: string;
+	url: string;
+}
+
+export interface Props {
+	remote: RemoteData;
+	removeLoader?: boolean;
+	token?: string;
+}
