--- packages/shared/components/RemoteMfe.types.ts
@@ -1,4 +1,0 @@
-export interface RemoteData {
-	name: string;
-	url: string;
-}
