+++ packages/shared/index.ts
@@ -1,4 +1,4 @@
-export { default as RemoteMfe } from './components/RemoteMfe';
-export { type RemoteData } from './components/RemoteMfe.types';
-export * from './utils';
+export { default as Counter } from './components/Counter/Counter';
+export { default as RemoteMfe } from './components/RemoreMfe/RemoteMfe';
+export { type RemoteData } from './components/RemoreMfe/RemoteMfe.types';
