+++ tsconfig.json
@@ -31,6 +31,7 @@
   "include": [
     "apps/host",
     "apps/remote",
+    "packages/shared",
     "@types/deno.d.ts",
     "@types/bun.d.ts",
     "types.d.ts"
